ugc/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
kafka-python==2.0.2
aiokafka==0.8.1
orjson==3.9.10
pyjwt[crypto]==2.8.0
pydantic-settings==2.1.0